refactor(A2-2564): hard code back link for appeal sub section back to overview page
