+++ packages/forms-web-app/src/controllers/selected-appeal/questionnaire-details/index.js
@@ -3,6 +3,7 @@ const { formatHeadlineData, formatQuestionnaireRows } = require('@pins/common');
 
 const { VIEW } = require('#lib/views');
 const { determineUser } = require('#lib/determine-user');
+const { getParentPathLink } = require('../../../lib/get-user-back-links');
 const {
 	formatQuestionnaireHeading,
 	formatTitleSuffix
@@ -30,6 +31,7 @@ exports.get = (layoutTemplate = 'layouts/no-banner-link/main.njk') => {
 		const appealNumber = req.params.appealNumber;
 		const trailingSlashRegex = /\/$/;
 		const userRouteUrl = req.originalUrl.replace(trailingSlashRegex, '').split('?')[0];
+		const backToAppealOverviewLink = getParentPathLink(userRouteUrl);
 
 		// determine user based on route to selected appeal
 		//i.e '/appeals/' = appellant | agent
@@ -95,6 +97,7 @@ exports.get = (layoutTemplate = 'layouts/no-banner-link/main.njk') => {
 
 		const viewContext = {
 			layoutTemplate,
+			backToAppealOverviewLink,
 			titleSuffix: formatTitleSuffix(userType),
 			mainHeading: formatQuestionnaireHeading(userType),
 			appeal: {
